requirements.txt
aiobotocore==1.3.3
aiohttp==3.7.4.post0
aioitertools==0.8.0
aiomysql==0.0.21
aioredis==1.3.1
aiosqlite==0.17.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
boto==2.49.0
boto3==1.17.0
botocore==1.20.106
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
ConfigArgParse==1.5.2
fastapi==0.68.0
fsspec==2021.7.0
gevent==21.8.0
geventhttpclient==1.4.4
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
locust==2.1.0
loguru==0.5.3
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
mysql-connector-python==8.0.26
newrelic==6.8.1.164
nltk==3.6.2
numpy==1.21.1
pandas==1.3.1
prometheus-client==0.11.0
prometheus-fastapi-instrumentator==5.7.1
protobuf==3.18.1
psutil==5.8.0
pyarrow==5.0.0
pydantic==1.8.2
PyMySQL==0.9.3
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.3
requests==2.26.0
rollbar==0.16.1
roundrobin==0.0.2
s3fs==2021.7.0
s3io==0.1.1
s3transfer==0.3.7
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.25
sqlalchemy2-stubs==0.0.2a17
sqlmodel==0.0.4
starlette==0.14.2
threadpoolctl==2.2.0
tqdm==4.62.0
typer==0.4.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
